pyproject.toml
[tool.poetry]
name = "cleaning-app"
version = "0.1.0"
description = ""
authors = ["leodpj <leodj@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.2.7"
djangorestframework = "^3.16.1"
mysqlclient = "^2.2.7"
python-decouple = "^3.8"
django-cors-headers = "^4.9.0"
serializers = "^0.2.4"
django-extensions = "^4.1"
djangorestframework-simplejwt = "^5.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
